+++ Environments/SharePoint-AllVersions/dsc/ConfigureDCVM.ps1
@@ -30,11 +30,11 @@
         #**********************************************************
         # Create AD domain
         #**********************************************************
-        WindowsFeature ADDS { Name = "AD-Domain-Services"; Ensure = "Present" }
-        WindowsFeature DNS  { Name = "DNS";                Ensure = "Present" }
+        WindowsFeature AddADDS { Name = "AD-Domain-Services"; Ensure = "Present" }
+        WindowsFeature AddDNS  { Name = "DNS";                Ensure = "Present" }
         DnsServerAddress SetDNS { Address = '127.0.0.1' ; InterfaceAlias = $InterfaceAlias; AddressFamily  = 'IPv4' }
         
-        ADDomain FirstDS
+        ADDomain CreateADForest
         {
             DomainName                    = $DomainFQDN
             Credential                    = $DomainCredsNetbios
@@ -42,23 +42,23 @@
             DatabasePath                  = "C:\NTDS"
             LogPath                       = "C:\NTDS"
             SysvolPath                    = "C:\SYSVOL"
-            DependsOn                     = "[DnsServerAddress]SetDNS"
+            DependsOn                     = "[DnsServerAddress]SetDNS", "[WindowsFeature]AddADDS"
         }
 
-        PendingReboot RebootAfterADDomainSignal
+        PendingReboot RebootOnSignalFromCreateADForest
         {
-            Name      = "RebootAfterADDomainSignal"
-            DependsOn = "[ADDomain]FirstDS"
+            Name      = "RebootOnSignalFromCreateADForest"
+            DependsOn = "[ADDomain]CreateADForest"
         }
 
-        WaitForADDomain ADDomainReady
+        WaitForADDomain WaitForDCReady
         {
             DomainName              = $DomainFQDN
             WaitTimeout             = 300
             RestartCount            = 3
             Credential              = $DomainCredsNetbios
             WaitForValidCredentials = $true
-            DependsOn               = "[PendingReboot]RebootAfterADDomainSignal"
+            DependsOn               = "[PendingReboot]RebootOnSignalFromCreateADForest"
         }
 
         #**********************************************************
@@ -71,16 +71,16 @@
             EmailAddress         = "$($Admincreds.UserName)@$DomainFQDN"
             PasswordNeverExpires = $true
             Ensure               = "Present"
-            DependsOn            = "[WaitForADDomain]ADDomainReady"
+            DependsOn            = "[WaitForADDomain]WaitForDCReady"
         }
 
         if ($ConfigureADFS -eq $true) {
             #**********************************************************
             # Configure AD CS
             #**********************************************************
-            WindowsFeature AddADCSFeature       { Name = "ADCS-Cert-Authority"; Ensure = "Present"; DependsOn = "[WaitForADDomain]ADDomainReady" }
+            WindowsFeature AddADCSFeature { Name = "ADCS-Cert-Authority"; Ensure = "Present"; DependsOn = "[WaitForADDomain]WaitForDCReady" }
 
-            ADCSCertificationAuthority ADCS
+            ADCSCertificationAuthority CreateADCSAuthority
             {
                 IsSingleInstance = "Yes"
                 CAType = "EnterpriseRootCA"
@@ -93,14 +93,14 @@
             {
                 CAServerFQDN = "$ComputerName.$DomainFQDN"
                 CARootName = "$DomainNetbiosName-$ComputerName-CA"
-                DependsOn = '[ADCSCertificationAuthority]ADCS'
+                DependsOn = '[ADCSCertificationAuthority]CreateADCSAuthority'
                 PsDscRunAsCredential = $DomainCredsNetbios
             }
 
             #**********************************************************
             # Configure AD FS
             #**********************************************************
-            CertReq ADFSSiteCert
+            CertReq GenerateADFSSiteCertificate
             {
                 CARootName                = "$DomainNetbiosName-$ComputerName-CA"
                 CAServerFQDN              = "$ComputerName.$DomainFQDN"
@@ -118,7 +118,7 @@
                 DependsOn = '[WaitForCertificateServices]WaitAfterADCSProvisioning'
             }
 
-            CertReq ADFSSigningCert
+            CertReq GenerateADFSSigningCertificate
             {
                 CARootName                = "$DomainNetbiosName-$ComputerName-CA"
                 CAServerFQDN              = "$ComputerName.$DomainFQDN"
@@ -135,7 +135,7 @@
                 DependsOn = '[WaitForCertificateServices]WaitAfterADCSProvisioning'
             }
 
-            CertReq ADFSDecryptionCert
+            CertReq GenerateADFSDecryptionCertificate
             {
                 CARootName                = "$DomainNetbiosName-$ComputerName-CA"
                 CAServerFQDN              = "$ComputerName.$DomainFQDN"
@@ -178,7 +178,7 @@
                     # If it returns $false, the SetScript block will run. If it returns $true, the SetScript block will not run.
                 return $false
                 }
-                DependsOn = "[CertReq]ADFSSiteCert", "[CertReq]ADFSSigningCert", "[CertReq]ADFSDecryptionCert"
+                DependsOn = "[CertReq]GenerateADFSSiteCertificate", "[CertReq]GenerateADFSSigningCertificate", "[CertReq]GenerateADFSDecryptionCertificate"
             }
 
             ADUser CreateAdfsSvcAccount
@@ -189,7 +189,7 @@
                 Ensure = "Present"
                 PasswordAuthentication = 'Negotiate'
                 PasswordNeverExpires = $true
-                DependsOn = "[CertReq]ADFSSiteCert", "[CertReq]ADFSSigningCert", "[CertReq]ADFSDecryptionCert"
+                DependsOn = "[CertReq]GenerateADFSSiteCertificate", "[CertReq]GenerateADFSSigningCertificate", "[CertReq]GenerateADFSDecryptionCertificate"
             }
 
             Group AddAdfsSvcAccountToDomainAdminsGroup
@@ -211,7 +211,7 @@
                 Target = $PrivateIP
                 Type = "ARecord"
                 Ensure = "Present"
-                DependsOn = "[WaitForADDomain]ADDomainReady"
+                DependsOn = "[WaitForADDomain]WaitForDCReady"
             }
 
             # https://docs.microsoft.com/en-us/windows-server/identity/ad-fs/deployment/configure-corporate-dns-for-the-federation-service-and-drs
@@ -222,7 +222,7 @@
                 Target = "$ComputerName.$DomainFQDN"
                 Type = "CName"
                 Ensure = "Present"
-                DependsOn = "[WaitForADDomain]ADDomainReady"
+                DependsOn = "[WaitForADDomain]WaitForDCReady"
             }
 
             cADFSFarm CreateADFSFarm
@@ -265,12 +265,14 @@ param = c.Value);
                 PsDscRunAsCredential = $DomainCredsNetbios
                 DependsOn = "[cADFSFarm]CreateADFSFarm"
             }
+            
+            WindowsFeature AddADCSManagementTools { Name = "RSAT-ADCS-Mgmt";     Ensure = "Present"; }
         }
 
-        WindowsFeature AddADFeature1 { Name = "RSAT-ADLDS";       Ensure = "Present"; }
-        WindowsFeature AddADFeature2 { Name = "RSAT-ADDS-Tools";  Ensure = "Present"; }
-        WindowsFeature DnsTools      { Name = "RSAT-DNS-Server";  Ensure = "Present"; }
-        WindowsFeature AddADCSManagementTools { Name = "RSAT-ADCS-Mgmt";      Ensure = "Present"; }
+        WindowsFeature AddADTools             { Name = "RSAT-AD-Tools";      Ensure = "Present"; }
+        WindowsFeature AddADPowerShell        { Name = "RSAT-AD-PowerShell"; Ensure = "Present"; }
+        WindowsFeature AddDnsTools            { Name = "RSAT-DNS-Server";    Ensure = "Present"; }
+        WindowsFeature AddADLDS               { Name = "RSAT-ADLDS";         Ensure = "Present"; }
     }
 }