+++ Environments/SharePoint-AllVersions/dsc/ConfigureSPVM.ps1
@@ -43,9 +43,9 @@ configuration ConfigureSPVM
         #**********************************************************
         # Initialization of VM - Do as much work as possible before waiting on AD domain to be available
         #**********************************************************
-        WindowsFeature ADTools  { Name = "RSAT-AD-Tools";      Ensure = "Present"; }
-        WindowsFeature ADPS     { Name = "RSAT-AD-PowerShell"; Ensure = "Present"; }
-        WindowsFeature DnsTools { Name = "RSAT-DNS-Server";    Ensure = "Present"; }
+        WindowsFeature AddADTools      { Name = "RSAT-AD-Tools";      Ensure = "Present"; }
+        WindowsFeature AddADPowerShell { Name = "RSAT-AD-PowerShell"; Ensure = "Present"; }
+        WindowsFeature AddDnsTools     { Name = "RSAT-DNS-Server";    Ensure = "Present"; }
         DnsServerAddress SetDNS { Address = $DNSServer; InterfaceAlias = $InterfaceAlias; AddressFamily  = 'IPv4' }
 
         # IIS cleanup
@@ -65,7 +65,6 @@ configuration ConfigureSPVM
             ValueData = "1"
             ValueType = "Dword"
             Ensure    = "Present"
-            DependsOn = "[PendingReboot]RebootOnComputerSignal"
         }
 
         # Properly enable TLS 1.2 as documented in https://docs.microsoft.com/en-us/azure/active-directory/manage-apps/application-proxy-add-on-premises-application
@@ -131,14 +130,13 @@ configuration ConfigureSPVM
             ServerName           = $SQLName
             Protocol             = "TCP"
             TcpPort              = 1433
-            #PsDscRunAsCredential = $DomainAdminCredsQualified
-            #DependsOn            = "[File]AccountsProvisioned"
         }
 
         #**********************************************************
         # Join AD forest
         #**********************************************************
-        WaitForADDomain DscForestWait
+        # If WaitForADDomain does not find the domain whtin "WaitTimeout" secs, it will signar a restart to DSC engine "RestartCount" times
+        WaitForADDomain WaitForDCReady
         {
             DomainName              = $DomainFQDN
             WaitTimeout             = 1200
@@ -149,63 +147,64 @@ configuration ConfigureSPVM
         }
         
         # WaitForADDomain sets reboot signal only if WaitForADDomain did not find domain within "WaitTimeout" secs
-        PendingReboot RebootOnWaitForADDomainSignal
+        PendingReboot RebootOnSignalFromWaitForDCReady
         {
-            Name             = "RebootOnWaitForADDomainSignal"
+            Name             = "RebootOnSignalFromWaitForDCReady"
             SkipCcmClientSDK = $true
-            DependsOn        = "[WaitForADDomain]DscForestWait"
+            DependsOn        = "[WaitForADDomain]WaitForDCReady"
         }
 
-        Computer DomainJoin
+        Computer JoinDomain
         {
             Name       = $ComputerName
             DomainName = $DomainFQDN
             Credential = $DomainAdminCredsQualified
-            DependsOn  = "[PendingReboot]RebootOnWaitForADDomainSignal"
+            DependsOn  = "[PendingReboot]RebootOnSignalFromWaitForDCReady"
         }
 
-        PendingReboot RebootOnComputerSignal
+        PendingReboot RebootOnSignalFromJoinDomain
         {
-            Name             = "RebootOnComputerSignal"
+            Name             = "RebootOnSignalFromJoinDomain"
             SkipCcmClientSDK = $true
-            DependsOn        = "[Computer]DomainJoin"
+            DependsOn        = "[Computer]JoinDomain"
         }
         
-        # This script might fix an issue that occured because VM did not reboot after it joined the domain.
-        # xScript CreateWSManSPNsIfNeeded
-        # {
-        #     SetScript =
-        #     {
-        #         # A few times, deployment failed because of this error:
-        #         # "The WinRM client cannot process the request. A computer policy does not allow the delegation of the user credentials to the target computer because the computer is not trusted."
-        #         # The root cause was that SPNs WSMAN/SP and WSMAN/sp.contoso.local were missing in computer account contoso\SP
-        #         # Those SPNs are created by WSMan when it (re)starts
-        #         # Restarting service causes an error, so creates SPNs manually instead
-        #         # Restart-Service winrm
+        # This script is still needed
+        xScript CreateWSManSPNsIfNeeded
+        {
+            SetScript =
+            {
+                # A few times, deployment failed because of this error:
+                # "The WinRM client cannot process the request. A computer policy does not allow the delegation of the user credentials to the target computer because the computer is not trusted."
+                # The root cause was that SPNs WSMAN/SP and WSMAN/sp.contoso.local were missing in computer account contoso\SP
+                # Those SPNs are created by WSMan when it (re)starts
+                # Restarting service causes an error, so creates SPNs manually instead
+                # Restart-Service winrm
 
-        #         # Create SPNs WSMAN/SP and WSMAN/sp.contoso.local
-        #         $domainFQDN = $using:DomainFQDN
-        #         $computerName = $using:ComputerName
-        #         Write-Verbose -Message "Adding SPNs 'WSMAN/$computerName' and 'WSMAN/$computerName.$domainFQDN' to computer '$computerName'"
-        #         setspn.exe -S "WSMAN/$computerName" "$computerName"
-        #         setspn.exe -S "WSMAN/$computerName.$domainFQDN" "$computerName"
-        #     }
-        #     GetScript = { return @{ "Result" = "false" } } # This block must return a hashtable. The hashtable must only contain one key Result and the value must be of type String.
-        #     TestScript = 
-        #     {
-        #         $computerName = $using:ComputerName
-        #         $samAccountName = "$computerName$"
-        #         if ((Get-ADComputer -Filter {(SamAccountName -eq $samAccountName)} -Property serviceprincipalname | Select-Object serviceprincipalname | Where-Object {$_.ServicePrincipalName -like "WSMAN/$computerName"}) -ne $null) {
-        #             # SPN is present
-        #             return $true
-        #         }
-        #         else {
-        #             # SPN is missing and must be created
-        #             return $false
-        #         }
-        #     }
-        #     DependsOn = "[PendingReboot]RebootOnComputerSignal"
-        # }
+                # Create SPNs WSMAN/SP and WSMAN/sp.contoso.local
+                $domainFQDN = $using:DomainFQDN
+                $computerName = $using:ComputerName
+                Write-Verbose -Message "Adding SPNs 'WSMAN/$computerName' and 'WSMAN/$computerName.$domainFQDN' to computer '$computerName'"
+                setspn.exe -S "WSMAN/$computerName" "$computerName"
+                setspn.exe -S "WSMAN/$computerName.$domainFQDN" "$computerName"
+            }
+            GetScript = { }
+            # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
+            TestScript = 
+            {
+                $computerName = $using:ComputerName
+                $samAccountName = "$computerName$"
+                if ((Get-ADComputer -Filter {(SamAccountName -eq $samAccountName)} -Property serviceprincipalname | Select-Object serviceprincipalname | Where-Object {$_.ServicePrincipalName -like "WSMAN/$computerName"}) -ne $null) {
+                    # SPN is present
+                    return $true
+                }
+                else {
+                    # SPN is missing and must be created
+                    return $false
+                }
+            }
+            DependsOn = "[PendingReboot]RebootOnSignalFromJoinDomain"
+        }
 
         #**********************************************************
         # Do SharePoint pre-reqs that require membership in AD domain
@@ -220,7 +219,7 @@ configuration ConfigureSPVM
             Type                 = "CName"
             Ensure               = "Present"
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[PendingReboot]RebootOnComputerSignal"
+            DependsOn            = "[PendingReboot]RebootOnSignalFromJoinDomain"
         }
 
         #**********************************************************
@@ -234,7 +233,7 @@ configuration ConfigureSPVM
             PasswordNeverExpires          = $true
             Ensure                        = "Present"
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn                     = "[PendingReboot]RebootOnComputerSignal"
+            DependsOn                     = "[PendingReboot]RebootOnSignalFromJoinDomain"
         }        
 
         ADUser CreateSParmAccount
@@ -245,7 +244,7 @@ configuration ConfigureSPVM
             PasswordNeverExpires          = $true
             Ensure                        = "Present"
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn                     = "[PendingReboot]RebootOnComputerSignal"
+            DependsOn                     = "[PendingReboot]RebootOnSignalFromJoinDomain"
         }
 
         Group AddSPSetupAccountToAdminGroup
@@ -255,7 +254,7 @@ configuration ConfigureSPVM
             MembersToInclude     = @("$($SPSetupCredsQualified.UserName)")
             Credential           = $DomainAdminCredsQualified
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[ADUser]CreateSPSetupAccount", "[ADUser]CreateSParmAccount"
+            DependsOn            = "[ADUser]CreateSPSetupAccount"
         }
 
         ADUser CreateSPAppPoolAccount
@@ -266,7 +265,7 @@ configuration ConfigureSPVM
             PasswordNeverExpires          = $true
             Ensure                        = "Present"
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn                     = "[PendingReboot]RebootOnComputerSignal"
+            DependsOn                     = "[PendingReboot]RebootOnSignalFromJoinDomain"
         }
 
         File AccountsProvisioned
@@ -276,7 +275,7 @@ configuration ConfigureSPVM
             Type                 = 'File'
             Force                = $true
             PsDscRunAsCredential = $SPSetupCredential
-            DependsOn            = "[Group]AddSPSetupAccountToAdminGroup", "[ADUser]CreateSParmAccount", "[ADUser]CreateSPAppPoolAccount"
+            DependsOn            = "[Group]AddSPSetupAccountToAdminGroup", "[ADUser]CreateSParmAccount", "[ADUser]CreateSPAppPoolAccount", "[xScript]CreateWSManSPNsIfNeeded"
         }
 
         xScript WaitForSQL
@@ -285,7 +284,7 @@ configuration ConfigureSPVM
             {
                 $retrySleep = 30
                 $server = $using:SQLAlias
-                $db="master"
+                $db = "master"
                 $retry = $true
                 while ($retry) {
                     $sqlConnection = New-Object System.Data.SqlClient.SqlConnection "Data Source=$server;Initial Catalog=$db;Integrated Security=True;Enlist=False;Connect Timeout=3"
@@ -302,7 +301,7 @@ configuration ConfigureSPVM
                 }
             }
             GetScript            = { return @{ "Result" = "false" } } # This block must return a hashtable. The hashtable must only contain one key Result and the value must be of type String.
-            TestScript           = { return $false } # If it returns $false, the SetScript block will run. If it returns $true, the SetScript block will not run.
+            TestScript           = { return $false } # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
             PsDscRunAsCredential = $DomainAdminCredsQualified
             DependsOn            = "[SqlAlias]AddSqlAlias", "[File]AccountsProvisioned"
         }
@@ -334,7 +333,7 @@ configuration ConfigureSPVM
             DependsOn            = "[SPFarm]CreateSPFarm"
         }
 
-        SPWebApplication MainWebApp
+        SPWebApplication CreateMainWebApp
         {
             Name                   = "SharePoint - 80"
             ApplicationPool        = "SharePoint - 80"
@@ -348,10 +347,10 @@ configuration ConfigureSPVM
             DependsOn              = "[SPFarm]CreateSPFarm"
         }
 
-        # Creating site collection in SharePoint 2019 fail: https://github.com/PowerShell/SharePointDsc/issues/990
+        # Creating site collection in SharePoint 2019 fails: https://github.com/PowerShell/SharePointDsc/issues/990
         # The workaround is to force a reboot before creating it
         if ($SharePointVersion -eq "2019") {
-            xScript SetRebootIfFirstTime
+            xScript ForceRebootBeforeCreatingSPSite
             {
                 # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
                 TestScript = {
@@ -363,14 +362,14 @@ configuration ConfigureSPVM
                 }
                 GetScript = { }
                 PsDscRunAsCredential = $SPSetupCredsQualified
-                DependsOn = "[SPWebApplication]MainWebApp"
+                DependsOn = "[SPWebApplication]CreateMainWebApp"
             }
 
             PendingReboot RebootBeforeCreatingSPSite
             {
                 Name             = "BeforeCreatingSPTrust"
                 SkipCcmClientSDK = $true
-                DependsOn        = "[xScript]SetRebootIfFirstTime"
+                DependsOn        = "[xScript]ForceRebootBeforeCreatingSPSite"
             }
         }
 
@@ -384,7 +383,7 @@ configuration ConfigureSPVM
                 SourcePath      = "$DCSetupPath"
                 DestinationPath = "$SetupPath\Certificates"
                 Credential      = $DomainAdminCredsQualified
-                DependsOn       = "[PendingReboot]RebootOnComputerSignal"
+                DependsOn       = "[PendingReboot]RebootOnSignalFromJoinDomain"
             }
 
             SPTrustedRootAuthority TrustRootCA
@@ -431,11 +430,11 @@ configuration ConfigureSPVM
                 }
                 GetScript            = { return @{ "Result" = "false" } } # This block must return a hashtable. The hashtable must only contain one key Result and the value must be of type String.
                 TestScript           = { return $false }
-                DependsOn            = "[PendingReboot]RebootOnComputerSignal"
+                DependsOn            = "[PendingReboot]RebootOnSignalFromJoinDomain"
                 PsDscRunAsCredential = $DomainAdminCredsQualified
             }
 
-            CertReq SPSSiteCert
+            CertReq GenerateMainWebAppCertificate
             {
                 CARootName             = "$DomainNetbiosName-$DCName-CA"
                 CAServerFQDN           = "$DCName.$DomainFQDN"
@@ -452,7 +451,7 @@ configuration ConfigureSPVM
                 DependsOn              = '[xScript]UpdateGPOToTrustRootCACert'
             }
 
-            SPWebApplicationExtension ExtendWebApp
+            SPWebApplicationExtension ExtendMainWebApp
             {
                 WebAppUrl              = "http://$SPTrustedSitesName/"
                 Name                   = "SharePoint - 443"
@@ -463,10 +462,10 @@ configuration ConfigureSPVM
                 Port                   = 443
                 Ensure                 = "Present"
                 PsDscRunAsCredential   = $SPSetupCredsQualified
-                DependsOn              = '[CertReq]SPSSiteCert'
+                DependsOn              = "[CertReq]GenerateMainWebAppCertificate", "[SPWebApplication]CreateMainWebApp"
             }
 
-            SPWebAppAuthentication ConfigureAuthentAllZones
+            SPWebAppAuthentication ConfigureMainWebAppAuthentication
             {
                 WebAppUrl = "http://$SPTrustedSitesName/"
                 Default = @(
@@ -482,7 +481,7 @@ configuration ConfigureSPVM
                     }
                 )
                 PsDscRunAsCredential = $SPSetupCredsQualified
-                DependsOn            = "[SPWebApplicationExtension]ExtendWebApp"
+                DependsOn            = "[SPWebApplicationExtension]ExtendMainWebApp"
             }
 
             xWebsite SetHTTPSCertificate
@@ -497,22 +496,23 @@ configuration ConfigureSPVM
                 }
                 Ensure               = "Present"
                 PsDscRunAsCredential = $DomainAdminCredsQualified
-                DependsOn            = "[SPWebAppAuthentication]ConfigureAuthentAllZones"
+                DependsOn            = "[SPWebAppAuthentication]ConfigureMainWebAppAuthentication"
             }
 
-            SPSite RootTeamSite
+            SPSite CreateRootSite
             {
                 Url                  = "http://$SPTrustedSitesName/"
                 OwnerAlias           = "i:0#.w|$DomainNetbiosName\$($DomainAdminCreds.UserName)"
                 SecondaryOwnerAlias  = "i:05.t|$DomainFQDN|$($DomainAdminCreds.UserName)@$DomainFQDN"
                 Name                 = "Team site"
                 Template             = "STS#0"
+                CreateDefaultGroups  = $true
                 PsDscRunAsCredential = $SPSetupCredsQualified
-                DependsOn            = "[SPWebApplication]MainWebApp"
+                DependsOn            = "[SPWebApplication]CreateMainWebApp"
             }
         }
         else {
-            SPWebAppAuthentication ConfigureAuthentDefaultZone
+            SPWebAppAuthentication ConfigureMainWebAppAuthentication
             {
                 WebAppUrl = "http://$SPTrustedSitesName/"
                 Default = @(
@@ -520,19 +520,20 @@ configuration ConfigureSPVM
                         AuthenticationMethod = "WindowsAuthentication"
                         WindowsAuthMethod    = "NTLM"
                     }
-                )                
+                )
                 PsDscRunAsCredential = $SPSetupCredsQualified
-                DependsOn            = "[SPWebApplication]MainWebApp"
+                DependsOn            = "[SPWebApplication]CreateMainWebApp"
             }
 
-            SPSite RootTeamSite
+            SPSite CreateRootSite
             {
                 Url                  = "http://$SPTrustedSitesName/"
                 OwnerAlias           = "i:0#.w|$DomainNetbiosName\$($DomainAdminCreds.UserName)"
                 Name                 = "Team site"
                 Template             = "STS#0"
+                CreateDefaultGroups  = $true
                 PsDscRunAsCredential = $SPSetupCredsQualified
-                DependsOn            = "[SPWebApplication]MainWebApp"
+                DependsOn            = "[SPWebApplication]CreateMainWebApp"
             }
         }
     }
@@ -582,7 +583,7 @@ $SQLAlias = "SQLAlias"
 $SharePointVersion = "2019"
 $ConfigureADFS = $false
 
-$outputPath = "C:\Packages\Plugins\Microsoft.Powershell.DSC\2.80.0.2\DSCWork\ConfigureSPVM.0\ConfigureSPVM"
+$outputPath = "C:\Packages\Plugins\Microsoft.Powershell.DSC\2.80.0.3\DSCWork\ConfigureSPVM.0\ConfigureSPVM"
 ConfigureSPVM -DomainAdminCreds $DomainAdminCreds -SPSetupCreds $SPSetupCreds -SPFarmCreds $SPFarmCreds -SPAppPoolCreds $SPAppPoolCreds -SPPassphraseCreds $SPPassphraseCreds -DNSServer $DNSServer -DomainFQDN $DomainFQDN -DCName $DCName -SQLName $SQLName -SQLAlias $SQLAlias -SharePointVersion $SharePointVersion -ConfigureADFS $ConfigureADFS -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath $outputPath
 Set-DscLocalConfigurationManager -Path $outputPath
 Start-DscConfiguration -Path $outputPath -Wait -Verbose -Force