+++ Environments/SharePoint-AllVersions/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## September 2020 update
 
+* Many modifications on DSC scripts to improve their reliability, readability and consistency
+* Create default SharePoint security groups on team site
 * Ensure compliance with policy CASG-DenyNSGRule100Allow
 
 ## August 2020-08-18 update